Move da:preview helper code into scripts/da
Only mapEigenProjectData stays in src (next to the indexer that uses it in
production). Mocha spec extended to pick up script tests.
